Check Start And End Times
Created the code to verify if the start and end times for the appointment are within the business hours.

Note: I also added a note telling the user what the business hours are within their time zone.

Side note: There is an issue with the formatting of the converted business hours time hint, but I don't plan on fixing it prior to submitting the PA.