Add Appointment v2
Added a check to see if the start date for the appointment to be added is overlaps with an end date in the database.

The user should be able to add an appointment if the times don't overlap.

Note: I also added a check for if the end date for the appointment to be added overlaps with a start date.