Update Appointment v2
Fixed the formatting issue with the start and end time by adding a "0" at the end if the time has 00 minutes.

The update form populates all fields with the appointment that was selected to update.

Need to fix issue with saving and not being able to save because the code flags for overlapping time for the appointment that you're updating.