Update Appointment COMPLETED
Completed the code for the update appointment form.

I ended up adding an AND condition to all of my methods which search for overlapping appointments to skip any appointments in the database with the same appointment id as the appointment being updated.

Then in the method that I add up the total appointments, I added a new condition to add +1 to the total and see if it's the same as the size of the list of appointments from the database.

The thought process behind this is that since I'm skipping an appointment with the same id as the one being updated, the total will be off by 1 so I have to add that back.